+++ CHANGELOG.md
@@ -17,7 +17,7 @@
 
 ### Fixed
 
-- Allow canceled `Scope.destroyAndWait()` calls to be retried.
+- Finish `Scope.destroyAndWait()` cleanup after caller cancellation.
 
 ### Security